requirements.txt
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
google-generativeai==0.8.3
streamlit==1.39.0
pytube==15.0.0
